requirements.txt
Django==4.0.6
django-cors-headers==3.10.0
djangorestframework==3.13.0